% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSF_basic_func.R
\name{getPost_normal1}
\alias{getPost_normal1}
\title{Basic Bayesian functions related to CSF simulation}
\usage{
getPost_normal1(dat)
}
\arguments{
\item{data}{input dataset}
}
\description{
Several basic working functions to generate posterior
sample for common data types. Currently, functions are
avaliable for data with normal, negative binomial,
binary distributions.

`getPost_normal1`: normal likelihood with non-informative prior
(mean ~ normal(0, 10000), sd ~ unif(0, 8))

`getPost_normal2`: normal likelihood with conjugated prior (normal inverse gamma)

`getPost_ngbn`: negative binomial likelihood with non-informative prior
(rate ~ unif(0, 8), p ~ beta(1, 1))

`getPost_binary`: Bernoulli likelihood with beta prior
}
