% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{getPower_binary_fisher}
\alias{getPower_binary_fisher}
\alias{getSampleSize_binary_fisher}
\title{Sample size calculation: binary endpoint with Fisher test}
\usage{
getPower_binary_fisher(
  total_sample_size,
  ss_ratio = 1,
  mean_AC,
  mean_PLB,
  drop_rate = 0.1,
  conf.level = 0.95
)

getSampleSize_binary_fisher(
  power,
  ss_ratio = 1,
  mean_AC,
  mean_PLB,
  drop_rate = 0.1,
  conf.level = 0.95
)
}
\arguments{
\item{total_sample_size}{total sample size of two arms}

\item{ss_ratio}{randomization ratio between two arms}

\item{mean_AC}{treatment effect of active arm}

\item{mean_PLB}{PLB effect}

\item{drop_rate}{early termination rate}

\item{conf.level}{confidence level}

\item{power}{power}
}
\value{
see description part
}
\description{
Sample size calculation for binary endpoint with Fisher test.

`getPower_binary_fisher()` provides power with given sample size and effect size.

`getSampleSize_Zstat()` provide total sample size needed to reach specified power and effect size.
}
\examples{
## get power for a 240 pts study with 10\% drop out rate,
## randomization ratio 1:1, 50\% and 30\% response rate for
## two arms separately, at 0.05 significance level

getPower_binary_fisher(240, 1, 0.5, 0.3, 0.1, 0.95)

## search sample size needed to reach 90\% power with same setup

# comment out due to time-consuming code
#getSampleSize_binary_fisher(0.9, 1, 0.5, 0.3, 0.1, 0.95)

}
