% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{getPower_count_ovdisp}
\alias{getPower_count_ovdisp}
\alias{getSampleSize_count_ovdisp}
\title{Sample size calculation: count endpoint with overdispersion}
\usage{
getPower_count_ovdisp(
  total_sample_size,
  ss_ratio = 1,
  rate_AC,
  rate_PLB,
  over_disp_rate = 0.2,
  drop_rate = 0.1,
  conf.level = 0.95,
  n_sim = 10000,
  n_core = 6,
  glm_model = "negativebinomial"
)

getSampleSize_count_ovdisp(
  power,
  ss_ratio = 1,
  rate_AC,
  rate_PLB,
  over_disp_rate = 0.2,
  drop_rate = 0.1,
  conf.level = 0.95,
  n_sim = 10000,
  n_core = 6,
  glm_model = "negativebinomial"
)
}
\arguments{
\item{total_sample_size}{total sample size of two arms}

\item{ss_ratio}{randomization ratio between two arms}

\item{rate_AC}{event rate of active arm}

\item{rate_PLB}{event rate of PLB arm}

\item{over_disp_rate}{overdispersion rate, var/mean = 1+rate}

\item{drop_rate}{early termination rate}

\item{conf.level}{confidence level}

\item{n_sim}{number of simulations}

\item{n_core}{number of cores used for parallel computation}

\item{glm_model}{specify which glm model used in simulation.
It can be c("negativebinomial", "quasipoisson").}

\item{power}{power}
}
\value{
see description part
}
\description{
Sample size calculation for count endpoint with overdispersion.
In the simulation, overdispersion data is simulated with
negative binomial. Generated data is fitted with quasipoisson glm.

`getPower_count_ovdisp()` provide power with given sample size

`getSampleSize_count_ovdisp` provide sample size with given power
}
\examples{
## comment out due to time consuming example
## and parallel computation limitation in CRAN

## get power for a 80 pts study with 10\% drop out rate,
## randomization ratio 1:1, 1 and 0.5 as the event rate for
## two arms separately, 20\% over-dispersive at 0.05 significance level

#getPower_count_ovdisp(80, 1, 1, 0.5, 0.2, 0.1, 0.95)

## search sample size needed to reach 90\% power with same setup

#getSampleSize_count_ovdisp(0.9, 1, 1, 0.5, 0.2, 0.1, 0.95)

}
