% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{getPower_Zstat}
\alias{getPower_Zstat}
\alias{getEffectSize_Zstat}
\alias{getSampleSize_Zstat}
\alias{getPower_Tstat}
\alias{getSampleSize_Tstat}
\title{Sample size calculation: continuous endpoint and Z/t statistic}
\usage{
getPower_Zstat(
  total_sample_size,
  ss_ratio = 1,
  mean_AC,
  mean_PLB,
  sd,
  drop_rate = 0.1,
  conf.level = 0.95
)

getEffectSize_Zstat(
  total_sample_size,
  ss_ratio = 1,
  sd,
  drop_rate = 0.1,
  conf.level = 0.95
)

getSampleSize_Zstat(
  power,
  ss_ratio = 1,
  mean_AC,
  mean_PLB,
  sd,
  drop_rate = 0.1,
  conf.level = 0.95
)

getPower_Tstat(
  total_sample_size,
  ss_ratio = 1,
  mean_AC,
  mean_PLB,
  sd,
  drop_rate = 0.1,
  conf.level = 0.95
)

getSampleSize_Tstat(
  power,
  ss_ratio = 1,
  mean_AC,
  mean_PLB,
  sd,
  drop_rate = 0.1,
  conf.level = 0.95
)
}
\arguments{
\item{total_sample_size}{total sample size of two arms}

\item{ss_ratio}{randomization ratio between two arms}

\item{mean_AC}{treatment effect of active arm}

\item{mean_PLB}{PLB effect}

\item{sd}{standard deviation}

\item{drop_rate}{early termination rate}

\item{conf.level}{confidence level}

\item{power}{power}
}
\value{
see description part
}
\description{
Sample size calculation for continous endpoint with Z/t statistic.

`getPower_Zstat()` and `getPower_Tstat()` provides power with given sample size and effect size.

`getEffectSize_Zstat()` provides minimal effect size can be detected (defined as statistical significant) with given sample size.

`getSampleSize_Zstat()` and `getSampleSize_Tstat` provide total sample size needed to reach specified power and effect size.
}
\examples{
## using Z-test as an example
## get power for a 60 pts study with 10\% drop out rate,
## randomization ratio 1:1, 5 and 3 mean for
## two arms separately with common sd 3, at 0.05 significance level

getPower_Zstat(60, 1, 5, 3, 3, 0.1, 0.95)

## search sample size needed to reach 90\% power with same setup

getSampleSize_Zstat(0.9, 1, 5, 3, 3, 0.1, 0.95)

## calculate the minimal effect size can be detected
## (that is statistical significant) with 60 pts study
getEffectSize_Zstat(60, 1, 3, 0.1, 0.95)

}
