% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSampleSize.R
\docType{package}
\name{simpleSampleSize}
\alias{simpleSampleSize}
\alias{simpleSampleSize-package}
\title{simpleSampleSize: R package for sample size calculation}
\description{
Easy and straight implementation of sample size calculation with R. This package covers variable types including binary, continuous, count, etc.
}
\author{
\itemize{
  \item *maintainer, author* Chaoran Hu \email{huchaoran.stat@gmail.com}
}
}
